SF-121/feat: 관광지 검색 화면 구현

+++ src/main/java/com/ssafy/faraway/domain/attraction/dto/res/AttractionGetResponseDto.java
@@ -14,8 +14,8 @@ public class AttractionGetResponseDto {
     private String zipcode;
     private String tel;
     private String firstImage;
-    private int latitude;
-    private int longitude;
+    private double latitude;
+    private double longitude;
 
 
 }
